@@ -1,113 +1,107 @@
 {
-    "name": "wdio-image-comparison-service",
-    "version": "5.0.3",
-    "description": "Image comparison / visual regression testing for WebdriverIO",
-    "main": "build/index.js",
-    "types": "./types/index.d.ts",
-    "scripts": {
-        "clean": "rimraf coverage build .tmp",
-        "compile": "babel lib/ --ignore **/*.spec.js -d build/",
-        "lint": "prettier --write **/*.js && eslint ./lib ./tests",
-        "test": "jest",
-        "test.local.init": "wdio ./tests/configs/wdio.local.init.conf.js",
-        "test.local.desktop": "wdio tests/configs/wdio.local.desktop.conf.js",
-        "test.local.appium": "wdio tests/configs/wdio.local.appium.conf.js",
-        "test.local.desktop.multi": "wdio tests/configs/wdio.local.desktop.multi.conf.js",
-        "test.local.dev.tools.init": "wdio ./tests/configs/wdio.local.dev.tools.init.conf.js",
-        "test.local.dev.tools.desktop": "wdio tests/configs/wdio.local.dev.tools.desktop.conf.js",
-        "test.unit.coverage": "jest --coverage ",
-        "test.saucelabs": "wdio ./tests/configs/wdio.saucelabs.conf.js",
-        "test.saucelabs.desktop": "SAUCE_ENV=desktop wdio ./tests/configs/wdio.saucelabs.conf.js",
-        "test.saucelabs.emu": "SAUCE_ENV=emu wdio ./tests/configs/wdio.saucelabs.conf.js",
-        "test.saucelabs.sims": "SAUCE_ENV=sims wdio ./tests/configs/wdio.saucelabs.conf.js",
-        "watch": "npm run compile -- --watch",
-        "release": "release-it --github.release",
-        "release.ci": "npm run release -- --ci --npm.skipChecks --no-git.requireCleanWorkingDir",
-        "release.patch": "npm run release -- patch",
-        "release.minor": "npm run release -- minor",
-        "release.major": "npm run release -- major",
-        "precompile": "npm run clean",
-        "prepublish": "npm run compile",
-        "prerelease": "npm run compile",
-        "prepare": "husky install"
-    },
-    "keywords": [
-        "webdriverio",
-        "visual",
-        "regression",
-        "image",
-        "comparison"
+  "name": "wdio-image-comparison-service",
+  "version": "5.0.3",
+  "description": "Image comparison / visual regression testing for WebdriverIO",
+  "main": "build/index.js",
+  "type": "module",
+  "types": "./types/index.d.ts",
+  "scripts": {
+    "clean": "rimraf coverage build .tmp",
+    "precompile": "npm run clean",
+    "compile": "tsc",
+    "test": "jest",
+    "test.eslint": "eslint --cache lib tests",
+    "test.local.init": "rimraf localBaseline && wdio ./tests/configs/wdio.local.init.conf.ts",
+    "test.local.desktop": "wdio tests/configs/wdio.local.desktop.conf.ts",
+    "test.local.appium": "wdio tests/configs/wdio.local.appium.conf.ts",
+    "test.local.desktop.multi": "wdio tests/configs/wdio.local.desktop.multi.conf.ts",
+    "test.local.dev.tools.init": "wdio ./tests/configs/wdio.local.dev.tools.init.conf.ts",
+    "test.local.dev.tools.desktop": "wdio tests/configs/wdio.local.dev.tools.desktop.conf.ts",
+    "test.unit.coverage": "jest --coverage ",
+    "test.saucelabs": "wdio ./tests/configs/wdio.saucelabs.conf.ts",
+    "test.saucelabs.desktop": "SAUCE_ENV=desktop wdio ./tests/configs/wdio.saucelabs.conf.ts",
+    "test.saucelabs.emu": "SAUCE_ENV=emu wdio ./tests/configs/wdio.saucelabs.conf.ts",
+    "test.saucelabs.sims": "SAUCE_ENV=sims wdio ./tests/configs/wdio.saucelabs.conf.ts",
+    "prerelease": "npm run compile",
+    "release": "release-it --github.release",
+    "release.ci": "npm run release -- --ci --npm.skipChecks --no-git.requireCleanWorkingDir",
+    "release.patch": "npm run release -- patch",
+    "release.minor": "npm run release -- minor",
+    "release.major": "npm run release -- major",
+    "watch": "npm run compile -- -w",
+    "wdio": "wdio run ./wdio.conf.ts",
+    "prepare": "husky install"
+  },
+  "keywords": [
+    "webdriverio",
+    "visual",
+    "regression",
+    "image",
+    "comparison"
+  ],
+  "author": "wswebcreation",
+  "license": "MIT",
+  "devDependencies": {
+    "@tsconfig/node18": "^18.2.2",
+    "@types/eslint": "^8.44.6",
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
+    "@typescript-eslint/utils": "^6.9.0",
+    "@wdio/cli": "^8.20.5",
+    "@wdio/devtools-service": "^8.20.5",
+    "@wdio/globals": "^8.20.5",
+    "@wdio/local-runner": "^8.20.5",
+    "@wdio/mocha-framework": "^8.20.3",
+    "@wdio/sauce-service": "^8.20.5",
+    "@wdio/shared-store-service": "^8.20.5",
+    "@wdio/spec-reporter": "^8.20.0",
+    "chromedriver": "^118.0.1",
+    "devtools": "^8.20.5",
+    "eslint": "^8.52.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-unicorn": "^48.0.1",
+    "eslint-plugin-wdio": "^8.20.0",
+    "husky": "^8.0.0",
+    "jest": "^29.7.0",
+    "release-it": "^16.2.1",
+    "rimraf": "^5.0.5",
+    "shared-store": "^4.1.2",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2",
+    "wdio-chromedriver-service": "^8.1.1"
+  },
+  "dependencies": {
+    "@wdio/logger": "^8.16.17",
+    "@wdio/types": "^8.20.0",
+    "webdriver-image-comparison": "^2.0.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/wswebcreation/wdio-image-comparison-service.git"
+  },
+  "jest": {
+    "testMatch": [
+      "<rootDir>/lib/tests/*.js"
     ],
-    "author": "wswebcreation",
-    "license": "MIT",
-    "devDependencies": {
-        "@babel/cli": "^7.21.0",
-        "@babel/core": "^7.21.4",
-        "@babel/eslint-parser": "^7.21.3",
-        "@babel/preset-env": "^7.21.4",
-        "@babel/register": "^7.21.0",
-        "@wdio/cli": "^7.31.1",
-        "@wdio/devtools-service": "^7.31.1",
-        "@wdio/jasmine-framework": "^7.30.2",
-        "@wdio/local-runner": "^7.31.1",
-        "@wdio/sauce-service": "^7.31.1",
-        "@wdio/shared-store-service": "^7.31.1",
-        "@wdio/spec-reporter": "^7.31.1",
-        "babel-jest": "^29.5.0",
-        "chromedriver": "^112.0.0",
-        "devtools": "^7.31.1",
-        "eslint": "^8.39.0",
-        "eslint-config-prettier": "^8.8.0",
-        "eslint-plugin-import": "^2.27.5",
-        "eslint-plugin-jasmine": "^4.1.3",
-        "eslint-plugin-wdio": "^7.25.3",
-        "fs-extra": "^10.1.0",
-        "husky": "^8.0.3",
-        "jest": "^29.5.0",
-        "lint-staged": "^13.2.1",
-        "prettier": "2.8.8",
-        "release-it": "^15.10.1",
-        "rimraf": "^3.0.2",
-        "shared-store": "^4.1.2",
-        "wdio-chromedriver-service": "^7.3.2",
-        "webdriverio": "^7.31.1"
+    "testPathIgnorePatterns": [
+      "<rootDir>/tests/",
+      "<rootDir>/node_modules/"
+    ],
+    "coverageDirectory": "./coverage/",
+    "collectCoverage": true,
+    "coverageThreshold": {
+      "global": {
+        "branches": 100,
+        "functions": 100,
+        "lines": 100,
+        "statements": 100
+      }
     },
-    "dependencies": {
-        "@wdio/logger": "^7.26.0",
-        "webdriver-image-comparison": "^1.0.3"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/wswebcreation/wdio-image-comparison-service.git"
-    },
-    "jest": {
-        "testMatch": [
-            "<rootDir>/lib/tests/*.js"
-        ],
-        "testPathIgnorePatterns": [
-            "<rootDir>/tests/",
-            "<rootDir>/node_modules/"
-        ],
-        "coverageDirectory": "./coverage/",
-        "collectCoverage": true,
-        "coverageThreshold": {
-            "global": {
-                "branches": 100,
-                "functions": 100,
-                "lines": 100,
-                "statements": 100
-            }
-        },
-        "testEnvironment": "node",
-        "coveragePathIgnorePatterns": [
-            "node_modules/",
-            "tests/",
-            "docs/"
-        ]
-    },
-    "lint-staged": {
-        "**/*": "prettier --write --ignore-unknown",
-        "*.js": "eslint --cache --fix",
-        "*.{js,css,md}": "prettier --write"
-    }
+    "testEnvironment": "node",
+    "coveragePathIgnorePatterns": [
+      "node_modules/",
+      "tests/",
+      "docs/"
+    ]
+  }
 }
