@@ -1,8 +1,11 @@
-const { join } = require('path')
+import { join } from 'node:path'
+import type { DeviceOrientation, ExtendedSauceLabsCapabilities } from '../types/types.ts'
 
-module.exports = function sauceAndroidEmus({ buildName }) {
+export function sauceAndroidEmus({ buildName }: { buildName: string }) {
     const mobileSpecs = join(process.cwd(), './tests/specs/mobile.spec.js')
-    const chromeDriverPhones = ['LANDSCAPE', 'PORTRAIT']
+    const chromeDriverPhones = (
+        ['LANDSCAPE', 'PORTRAIT'] as DeviceOrientation[]
+    )
         .map((orientation) =>
             ['8.1', '9.0', '10.0', '11.0', '12.0'].map((platformVersion) =>
                 createCaps({
@@ -21,7 +24,9 @@ module.exports = function sauceAndroidEmus({ buildName }) {
             )
         )
         .flat(1)
-    const nativeWebScreenshotPhones = ['LANDSCAPE', 'PORTRAIT']
+    const nativeWebScreenshotPhones = (
+        ['LANDSCAPE', 'PORTRAIT'] as DeviceOrientation[]
+    )
         .map((orientation) =>
             ['8.1', '9.0', '10.0', '11.0', '12.0'].map((platformVersion) =>
                 createCaps({
@@ -41,7 +46,9 @@ module.exports = function sauceAndroidEmus({ buildName }) {
             )
         )
         .flat(1)
-    const chromeDriverTablets = ['LANDSCAPE', 'PORTRAIT']
+    const chromeDriverTablets = (
+        ['LANDSCAPE', 'PORTRAIT'] as DeviceOrientation[]
+    )
         .map((orientation) =>
             ['8.1', '11.0', '12.0'].map((platformVersion) => {
                 let tabletType = ''
@@ -75,7 +82,7 @@ module.exports = function sauceAndroidEmus({ buildName }) {
         )
         .flat(1)
     // // There is no Android 10 for Tablets
-    // const nativeWebScreenshotTablets = ['LANDSCAPE', 'PORTRAIT']
+    // const nativeWebScreenshotTablets = (['LANDSCAPE', 'PORTRAIT'] as DeviceOrientation[])
     //     .map((orientation) =>
     //         ['8.1', '11.0', '12.0'].map((platformVersion) => {
     //             let tabletType = ''
@@ -136,7 +143,28 @@ function createCaps({
     orientation,
     platformVersion,
     sauceOptions,
-}) {
+}: {
+    wdioIcsCommands?: string[];
+    deviceName: string;
+    mobileSpecs: string;
+    nativeWebScreenshot?: boolean;
+    orientation: string;
+    platformVersion: string;
+    sauceOptions: ExtendedSauceLabsCapabilities;
+}): {
+    browserName: string;
+    platformName: string;
+    'appium:deviceName': string;
+    'appium:platformVersion': string;
+    'appium:orientation': string;
+    'appium:automationName': string;
+    'wdio-ics:options': {
+        logName: string;
+        commands: string[];
+    };
+    'sauce:options': ExtendedSauceLabsCapabilities;
+    specs: string[];
+} {
     const driverScreenshotType = nativeWebScreenshot
         ? 'NativeWebScreenshot'
         : 'ChromeDriver'
